# api/swagger.yaml
openapi: 3.0.3
info:
  title: NeuroBus
  version: 1.0.0
servers:
  - url: https://14-bit.ru/api/
paths:
...
  /bus_station:
    get:
      tags:
        - BUS_STATION
      responses:
        '200':
          description: ok
    post:
      tags:
        - BUS_STATION
      parameters:
        - name: longitude
          in: query
          required: true
          schema:
            type:
              number
        - name: latitude
          in: query
          required: true
          schema:
            type:
              number
        - name: name
          in: query
          required: true
          schema:
            type:
              string
      responses:
        '200':
          description: ok
...
  /bus:
    get:
      tags:
        - BUS
      responses:
        '200':
          description: ok
    post:
      tags:
        - BUS
      parameters:
        - name: plateNumber
          in: query
          required: true
          schema:
            type:
              string
        - name: modelName
          in: query
          required: true
          schema:
            type:
              string
        - name: serialNumber
          in: query
          required: true
          schema:
            type:
              string
      responses:
        '200':
          description: ok
...
  /route:
    post:
      tags:
        - ROUTE
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type:
              string
      responses:
        '200':
          description: ok
    get:
      tags:
        - ROUTE
      responses:
        '200':
          description: ok
...
  /route/{routeId}:
    get:
      tags:
        - ROUTE
      description:
        Получить маршрут с его расписанием и остановками
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type:
              integer
      responses:
        '200':
          description: ok
...